feature: Change nested attr access format and add `callattr`
Before we had this weird sequence format for nested access.
This commit pulls all of this functionality out to a mixin class
and changes it to just a string of what you would have written
in code, ie 'attr["item"].subattr' etc, and adds `callattr` so
you can call nested methods on remote objects.