feature: RemoteComponent.update()
If something has changed on the server with the structure of the instrument,
like new or different parameters, call update() to have this reflected in the
RemoteInstrument. If just the component attribute values have changed this is
not necessary, but if component existence has changed, or which attributes
exist on the component has changed, you must manually call this.