feat: Add multiple parameter sweeps
Allows to sweep multiple parameters, pretty much with the same syntax
and some boiler plate.
Idea is to allow set a list of values, and then either save
the index as the set_point or an aggregated value if an aggregator
function is provided. The individual sets are also saved
as separate arrays, but they look like measurements arrays for
now.

See #111 and  #162.