driver: ziuhfli (#466)
* feat: Add a driver for ZI UHF-LI

First commit providing a driver with a working sweeper. Needs
refactorising the sweeper into a MultiParameter

* fix: Change the sweep into a MultiParameter

Change the sweep from a normal instrument method into a MultiParameter

Breaking Changes: execute_sweep no longer exists. All sweep data goes
through the Sweep parameter.

* fix: Tidy code

Tidy code and remove whitespace

* feat: Add example notebook

Add an example notebook explaining how to use the sweeper.

* fix: Correct multiple dict keys

Correct the multiply-defined dict keys of the demodulator trigger.

* updating ZIUHFLI.py on dummy setup

* intermediate commit

* testing"
"

* feat: Add parameters

Finish Sigout parameters and start adding scope parameters.

* Intermediate commit

To be cleaned later

* feat: Add scope get function

Add a get function to the scope multiparameter and document this in the
notebook.

* fix: Update setpoints with delay

Update the scope setpoints with reference and delay values.

* fix: Adjust setpoints

Adjust the setpoints of the scope to match the expectations of the loop.

* fix: pep8

Break lines and trim whitespace.

* fix: Add scope instance to instrument

Add a scope instance to the __init__ of the instrument.

* docs: Fix requested changes

Add requested docs