fix: #564 visa_handle.clear causes some instruments to freeze (#565)
* fix: #564 visa_handle.clear causes some instruments to freeze

removes visa_handle.clear when instrument is opened. This does
nothing anyway as buffers for newly opened instruments

* fix: #564 visa_handle.clear handle for serial instruments

Clear buffers for serial instruments using the flush method which is
only available for serial instruments, but does not perform additional
actions that may lock up the serial port.
This leaves the clear command in place for other instrument types that
do not support flush.