Feat: Slack bot integration (#427)
* Added slack bot

* Added latest_dataset/latest_plot

* refactor: moved slack.py to utils

* feat: upgraded slack, added active_loop, active_data_set

The new version of slack is a thread, and so does not need a ipython
widget. Furthermore, the active_loop and active_data_set functions
have been added, since slack needs access to the active loop and data_set.
active_loop() will point to the active loop. this is set when calling
loop.run(set_active=True), where set_active is True by default.
In some cases a measurement can itself consist of measurements.
For instance, a complex parameter could perform an entire measurement.
To avoid that the active_loop is overwritten, the kwarg set_active=False
can be added.

* added slacker to extra packages

* Removed .idea from PyCharm

* fix: Added catching of ValueError